Find an initial home for the refactor
