+++ lib/realms.rb
@@ -3,6 +3,7 @@ require "active_support/all"
 require "brainguy"
 require "realms/choices"
 require "realms/effects"
+require "realms/abilities"
 require "realms/actions"
 require "realms/cards"
 require "realms/game"
