+++ lib/realms/abilities.rb
@@ -1,0 +1,7 @@
+require "realms/abilities/builder"
+require "realms/abilities/definition"
+
+module Realms
+  module Abilities
+  end
+end
