+++ lib/realms/abilities/declarations.rb
@@ -1,0 +1,10 @@
+require "realms/abilities/declarations/sequence"
+require "realms/abilities/declarations/modal"
+require "realms/abilities/declarations/effect"
+
+module Realms
+  module Abilities
+    module Declarations
+    end
+  end
+end
