--- lib/realms/refactor.rb
@@ -1,9 +1,0 @@
-
-Dir[File.expand_path('../refactor/*.rb', __FILE__)].each do |file|
-  require file
-end
-
-module Realms
-  module Refactor
-  end
-end
